requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
openai>=1.0.0
anthropic>=0.3.0
google-cloud-aiplatform>=1.35.0
cohere>=4.0.0
networkx>=3.1
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
